requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
frozenlist==1.3.3
google-auth==2.17.1
google-auth-oauthlib==1.0.0
idna==3.4
langchain[llms]
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.7
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tenacity==8.2.2
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
python-dotenv
gunicorn
uvicorn
openai
faiss-cpu
unstructured
slack-bolt
supabase
streamlit
tiktoken
tabulate
pdf2image
pytesseract
google-search-results
pypdf
pycryptodome
streamlit_chat
python-magic-bin
psycopg2
langchain_experimental
scikit-learn
ratelimit